% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse_auxiliar.R
\name{block_sample}
\alias{block_sample}
\title{MSE helper functions}
\usage{
block_sample(domains, in_samp, smp_data, dom_name, pop_data, gb_res)
}
\arguments{
\item{domains}{domain}

\item{in_samp}{in_samp}

\item{smp_data}{sample data}

\item{dom_name}{domain name}

\item{pop_data}{population data}

\item{gb_res}{gb_res}
}
\description{
MSE helper functions
}
