% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{coef.MEGB}
\alias{coef.MEGB}
\title{Default Coef Method for MEGB Class}
\usage{
\method{coef}{MEGB}(megb_object, ...)
}
\arguments{
\item{megb_object}{object of class MEGB containing the boosting model}

\item{...}{additional arguments}
}
\value{
List of random effects coefficients and feature importance
}
\description{
Default Coef Method for MEGB Class
}
