% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{data_info}
\alias{data_info}
\title{information on datasets}
\usage{
data_info(dom_name, pop, smp)
}
\arguments{
\item{dom_name}{domain name}

\item{pop}{population dataset}

\item{smp}{sample dataset}
}
\value{
list of information on datasets
}
\description{
information on datasets
}
