% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_megb.R
\name{eval_params}
\alias{eval_params}
\title{This function evaluates the performance of a Gradient Boosting model
using the provided set of hyperparameters. It fits the model and
computes the best evaluation score based on the engine type (XGBoost,
LightGBM, or CatBoost).}
\usage{
eval_params(params, gbm_engine, features_train, response_train)
}
\arguments{
\item{params}{A list of hyperparameters for the Gradient Boosting model.}

\item{gbm_engine}{The engine used for the Gradient Boosting model. It can be
one of "xgboost", "lightgbm", or "catboost".}

\item{features_train}{A data frame or matrix containing the training features.}

\item{response_train}{A numeric vector containing the training response variable.}
}
\value{
A list containing:
\item{Score}{The best evaluation score (e.g., RMSE) of the model.}
\item{params}{The hyperparameters used in the model.}
\item{best_iter}{The best iteration (round) of the model during training.}
}
\description{
This function evaluates the performance of a Gradient Boosting model
using the provided set of hyperparameters. It fits the model and
computes the best evaluation score based on the engine type (XGBoost,
LightGBM, or CatBoost).
}
