% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_predict.R
\name{gbm_predict}
\alias{gbm_predict}
\title{MEGB helper function}
\usage{
gbm_predict(model, smp_data, pop_data, Y, dom_name, gbm_engine, cov_names)
}
\arguments{
\item{model}{model object}

\item{smp_data}{dataframe, sample data}

\item{pop_data}{dataframe, population data}

\item{Y}{target variable}

\item{dom_name}{factor, domain name}

\item{gbm_engine}{character, name of gbm engine to use}

\item{cov_names}{names of covariates}
}
\value{
list of unit predictions
}
\description{
MEGB helper function
}
