% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_importance.R
\name{get_feature_importance}
\alias{get_feature_importance}
\title{get feature importance - MEGB helper function}
\usage{
get_feature_importance(engine, model, features_train = NULL, train_pool = NULL)
}
\arguments{
\item{engine}{character, used engine}

\item{model}{list, megb model}

\item{features_train}{covariates used for training, default is NULL}

\item{train_pool}{train pool for catboost feature importance estimation, default is NULL}
}
\description{
get feature importance - MEGB helper function
}
