% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/megb.R
\name{input_checks_megb}
\alias{input_checks_megb}
\title{This function defines input checks for Mixed Effect Gradient Boosting,
which combines Gradient Boosting with Linear Mixed Models (LMM).}
\usage{
input_checks_megb(
  Y,
  X,
  dom_name,
  smp_data,
  pop_data,
  gradient_params,
  na.rm,
  seed,
  bootstrap_cores,
  mse,
  B,
  gbm_engine
)
}
\arguments{
\item{Y}{numeric, target variable}

\item{X}{data.frame, Covariates}

\item{dom_name}{factor, name of the domain variable}

\item{smp_data}{data.frame, sample data}

\item{pop_data}{data.frame, population data}

\item{gradient_params}{list, optional list of hyperparameters passed to the selected GB engine and required to match the specification of gbm_engine.}

\item{na.rm}{logical, should NAs be removed?}

\item{seed}{integer, seed to reproduce results}

\item{bootstrap_cores}{integer, number of cores to use for bootstrap}

\item{mse}{logical, should MSE be estimated?}

\item{B}{integer, number of bootstrap samples for MSE}

\item{gbm_engine}{character, choice of Gradient Boosting implementation ("xgboost", "lightgbm", "catboost")}
}
\value{
logical to indicate if tests were passed or not
}
\description{
This function defines input checks for Mixed Effect Gradient Boosting,
which combines Gradient Boosting with Linear Mixed Models (LMM).
}
