% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{load_megb}
\alias{load_megb}
\title{Load objects of class MEGB}
\usage{
load_megb(file, ...)
}
\arguments{
\item{file}{file to load object from}

\item{...}{additional arguments}
}
\value{
an R object
}
\description{
Load objects of class MEGB
}
