% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/megb.R
\name{megb}
\alias{megb}
\title{This function defines Mixed Effect Gradient Boosting,
which combines Gradient Boosting with Linear Mixed Models (LMM).}
\usage{
megb(
  Y,
  X,
  dom_name,
  smp_data,
  pop_data,
  gradient_params = NULL,
  na.rm = TRUE,
  seed,
  mse = FALSE,
  B = 100,
  bootstrap_cores = 0,
  gbm_engine = "xgboost",
  ...
)
}
\arguments{
\item{Y}{numeric, target variable}

\item{X}{data.frame, Covariates}

\item{dom_name}{factor, name of the domain variable}

\item{smp_data}{data.frame, sample data}

\item{pop_data}{data.frame, population data}

\item{gradient_params}{list, optional list to define Gradient Boosting parameters. If used, has to use the parameter names of the used GB engine.}

\item{na.rm}{logical, should NAs be removed?}

\item{seed}{integer, seed to reproduce results}

\item{mse}{logical, should MSE be estimated?}

\item{B}{numeric, number of bootstrap samples for MSE}

\item{bootstrap_cores}{integer, number of cores to use}

\item{gbm_engine}{character, choice of Gradient Boosting implementation ("xgboost", "lightgbm", "catboost")}

\item{...}{additional arguments}
}
\value{
list of class MEGB with results:
\itemize{
\item Indicators, containing the average domain predictions
\item gbmodel, containing the model details
\item inp_smp_data, containing the input data
\item unit_preds_all, containing the unit-level predictions (population)
\item unit_preds_all, containing the unit-level predictions (sample)
\item MSE_Estimates, containing te estimates MSE values
\item time_gradient, containing the runtime
\item gradient_params, containing the used Gradient Boosting parameters
}
}
\description{
This function defines Mixed Effect Gradient Boosting,
which combines Gradient Boosting with Linear Mixed Models (LMM).
}
\examples{
\donttest{
# Loading data - population and sample data
data("population_data")
data("sample_data")

# Example 1: with MSE = TRUE

params <- list(eta = 0.1,      # learning rate
               nrounds = 100   # number of iterations
               )

result <- megb(Y = sample_data$y,
               X = sample_data[, 2:6],
               dom_name = "domain",
               smp_data = sample_data,
               pop_data = population_data,
               gradient_params = params,
               na.rm = TRUE,
               seed = 1,
               mse = TRUE,
               B = 2,
               bootstrap_cores = 0,
               gbm_engine = "xgboost"
               )
}
}
