% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse_megb.R
\name{mse_megb}
\alias{mse_megb}
\title{MSE Estimation for Mixed Effect Gradient Boosting}
\usage{
mse_megb(
  Y,
  X,
  dom_name,
  smp_data,
  model,
  error_sd,
  pop_data,
  B,
  initial_random_effects,
  ErrorTolerance,
  MaxIterations,
  cov_names,
  gradient_params = list(),
  formula_random_effects,
  bootstrap_cores,
  seed,
  gbm_engine,
  unit_pred_smp,
  unit_preds,
  ...
)
}
\arguments{
\item{Y}{numeric, target variable}

\item{X}{data.frame, Covariates}

\item{dom_name}{factor, name of domain variable}

\item{smp_data}{data.frame, sample data}

\item{model}{list, trained MEGB model}

\item{error_sd}{numeric, standard deviation of error}

\item{pop_data}{data.frame, population data}

\item{B}{integer, number of bootstrap samples for MSE}

\item{initial_random_effects}{numeric, initial value for Random Effects}

\item{ErrorTolerance}{numeric, tolerance for EM-algorithm}

\item{MaxIterations}{integer, maximal number of iterations for the EM-algorithm}

\item{cov_names}{character, names of covariates}

\item{gradient_params}{list, optional list with Gradient Boosting parameters}

\item{formula_random_effects}{formula, formula for Random Effects}

\item{bootstrap_cores}{integer, number of cores to use for bootstrap}

\item{seed}{integer, choose seed}

\item{...}{additional arguments}
}
\value{
list of results of MSE estimates
}
\description{
This function calculates the Mean Squared Error (MSE) Estimates for a
Mixed Effect Gradient Boosting model while using bootstrap samples.
}
