% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/megb.R
\docType{data}
\name{params}
\alias{params}
\title{MEGB}
\format{
An object of class \code{list} of length 8.
}
\usage{
params
}
\value{
list of class MEGB with results
}
\description{
Gradient Boosting parameter,
}
\keyword{datasets}
