% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_megb.R
\name{plot.MEGB}
\alias{plot.MEGB}
\title{Plot Random Effects, Residuals, Train vs. Validation Loss, and Feature Importance}
\usage{
\method{plot}{MEGB}(megb_object)
}
\arguments{
\item{megb_object}{object of class MEGB which contains the boosting model}

\item{...}{additional arguments}
}
\value{
series of four plots, i.e. a QQ-Plot of the Random Effects, a QQPlot of the Residuals, plot containing the loss for train and validation data, a feature importance plot
}
\description{
Plot Random Effects, Residuals, Train vs. Validation Loss, and Feature Importance
}
