% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_data.R
\docType{data}
\name{population_data}
\alias{population_data}
\title{Simulated population data}
\format{
\subsection{\code{population_data}}{

A data frame with 50000 observations and 7 variables:
\describe{
\item{y}{numeric; the target variable.}
\item{x1}{numeric; first covariate}
\item{x2}{numeric; second covariate}
\item{x3}{numeric; third covariate}
\item{x4}{numeric; fourth covariate}
\item{x5}{numeric; fifth covariate}
\item{domain}{factor; domain}
}
}
}
\usage{
population_data
}
\description{
The data set includes domains, covariates and target variable y
}
\keyword{datasets}
