% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.MEGB}
\alias{print.MEGB}
\title{S3 print-method for MEGB-objects}
\usage{
\method{print}{MEGB}(megb_object, ...)
}
\arguments{
\item{megb_object}{object of class MEGB containing the boosting model}

\item{...}{additional arguments}
}
\value{
summary of model information
}
\description{
S3 print-method for MEGB-objects
}
