% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{save_megb}
\alias{save_megb}
\title{Save objects of class MEGB}
\usage{
save_megb(megb_object, file, ...)
}
\arguments{
\item{megb_object}{object of class MEGB containing the boosting model}

\item{file}{file to save object to}

\item{...}{additional arguments}
}
\value{
NULL invisibly
}
\description{
Save objects of class MEGB
}
