% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_gb_model.R
\name{train_gbmodel}
\alias{train_gbmodel}
\title{train GB model - MEGB helper function}
\usage{
train_gbmodel(
  model_type,
  features_train,
  response_train,
  params,
  cat_idx0 = integer(0),
  ...
)
}
\arguments{
\item{model_type}{character, indicating which gbm_engine was used}

\item{features_train}{covariates used for training}

\item{response_train}{response var used for training}

\item{params}{parameters to use for models}

\item{cat_idx0}{index for catboost, default is 0}

\item{...}{additional arguments}
}
\value{
list containing boosting information
}
\description{
train GB model - MEGB helper function
}
